chore(cubestore): Small refactoring of workers startup
